Splitting hp_tuning
